excellent progress
ChatGPT took the wheel and made excellent progress. Unfortunately there was quite a bit of copying and pasting involved. But I'm glad I was able to fully understand everything I'm putting into my code. Unfortunately I'm dealing with things more complicated than I'm currently capable of doing myself, and so I couldn't write the code myself as much as I'd have liked to. Still, it's reassuring to be able to understand the code I'm working with even though it's the first time I'm being introduced to so many new concepts.